moved campaign instructions to own file
